refactor: improve naming of logic for loading repositories and running a script
for #219
